requirements.txt
flask==2.3.2
torch==2.6.0
torchvision==0.21.0
transformers==4.31.0
tensorflow==2.19.0
Pillow==9.5.0
numpy==1.24.3
scikit-learn==1.2.2
gunicorn==20.1.0  # For production